Cast double value to int on bloc example

+++ example/lib/bloc_implementation/bloc_implementation_page.dart
@@ -167,7 +167,7 @@ class _BlocImplementationPageState extends State<BlocImplementationPage> {
                                     onChanged: (value) => context
                                         .read<FormXCubit<String>>()
                                         .updateAndValidateField(
-                                          value,
+                                          value.toInt(),
                                           'salaryExpectation',
                                         ),
                                   ),
